API Listing and treatmentsCode

+++ app/Http/Controllers/Api/PractitionerController.php
@@ -12,15 +12,30 @@ use Str;
 
 class PractitionerController extends Controller
 {
+
+
     /**
      * Display a listing of the resource.
      */
     public function index()
     {
-        $practitioners = Practitioner::all();
+        $user = Auth::user();
+
+        if (!$user) {
+            return response()->json([
+                'message' => 'User not authenticated'
+            ], 401);
+        }
+        $practitioners = Practitioner::where('user_id', $user->id)
+            ->orderBy('created_at', 'desc')
+            ->get();
 
         if ($practitioners->isEmpty()) {
             return response()->json(['message' => 'No practitioners found'], 404);
+        }
+
+        if ($practitioners->user_id != $user->id) {
+            return response()->json(['message' => 'You are not authorized to perform this action.'], 403);
         }
 
         return response()->json($practitioners);
@@ -47,6 +62,10 @@ class PractitionerController extends Controller
     public function update(PractitionerUpdateRequest $request, string $id)
     {
         $practitioner = Practitioner::findOrFail($id);
+        $user = Auth::user();
+        if ($practitioner->user_id != $user->id) {
+            return response()->json(['message' => 'You are not authorized to perform this action.'], 403);
+        }
         $practitioner->name = $request->name;
         $practitioner->qualification = $request->qualification;
         $practitioner->certificate = $request->certificate;
@@ -60,7 +79,11 @@ class PractitionerController extends Controller
     public function destroy(string $id)
     {
         $practitioner = Practitioner::findOrFail($id);
+        $user = Auth::user();
 
+        if ($practitioner->user_id != $user->id) {
+            return response()->json(['message' => 'You are not authorized to perform this action.'], 403);
+        }
         // Delete the practitioner
         $practitioner->delete();
